Moved pcm conversion to AudioConverter
