+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/audio/AudioConverter.kt
@@ -19,7 +19,14 @@
 package org.wycliffeassociates.otter.common.domain.audio
 
 import io.reactivex.Completable
+import org.wycliffeassociates.otter.common.audio.AudioFileReader
+import org.wycliffeassociates.otter.common.audio.DEFAULT_BITS_PER_SAMPLE
+import org.wycliffeassociates.otter.common.audio.DEFAULT_CHANNELS
+import org.wycliffeassociates.otter.common.audio.DEFAULT_SAMPLE_RATE
+import org.wycliffeassociates.otter.common.audio.pcm.PcmFile
+import org.wycliffeassociates.otter.common.audio.wav.WavFile
 import java.io.File
+import java.io.OutputStream
 import javax.inject.Inject
 import de.sciss.jump3r.Main as jump3r
 
@@ -38,5 +45,42 @@ class AudioConverter @Inject constructor() {
             )
             jump3r().run(args)
         }
+    }
+
+    fun wavToPcm(wavFile: File, pcmFile: File) {
+        val wavReader = WavFile(wavFile).reader().also { it.open() }
+        val pcmWriter = PcmFile(pcmFile).writer(append = false)
+
+        val buffer = ByteArray(10240)
+        while (wavReader.hasRemaining()) {
+            val written = wavReader.getPcmBuffer(buffer)
+            pcmWriter.write(buffer, 0, written)
+        }
+
+        cleanup(wavReader, pcmWriter)
+    }
+
+    fun pcmToWav(pcmFile: File, wavFile: File) {
+        val pcmReader = PcmFile(pcmFile).reader().also { it.open() }
+        val wavWriter = WavFile(
+            wavFile,
+            DEFAULT_CHANNELS,
+            DEFAULT_SAMPLE_RATE,
+            DEFAULT_BITS_PER_SAMPLE
+        ).writer(append = false)
+
+        val buffer = ByteArray(10240)
+        while (pcmReader.hasRemaining()) {
+            val written = pcmReader.getPcmBuffer(buffer)
+            wavWriter.write(buffer, 0, written)
+        }
+
+        cleanup(pcmReader, wavWriter)
+    }
+
+    private fun cleanup(reader: AudioFileReader, writer: OutputStream) {
+        reader.release()
+        writer.flush()
+        writer.close()
     }
 }
